tutorial/requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
decorator==4.3.0
Django==2.1.3
django-cors-headers==2.4.0
django-extensions==2.1.3
djangorestframework==3.9.0
httpie==1.0.0
idna==2.7
ipython==7.1.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10
MarkupSafe==1.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
